pyproject.toml
[project]
name = "gen-ai-first-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.0.8",
    "langchain-openai>=1.0.3",
    "python-dotenv>=1.2.1",
]
